+++ src/components/mainMenu/Menu.tsx
@@ -83,7 +83,7 @@ const Menu = ({ setIsNewGame }: Props) => {
 
   return (
     <>
-      <div className="fixed top-0 left-0 h-full w-full flex flex-col justify-between bg-cute bg-no-repeat bg-center p-5 items-center">
+      <div className="flex flex-col justify-between bg-cute bg-no-repeat bg-center bg-cover p-5 items-center h-screen w-screen">
         <div className="flex flex-col gap-5">
           <div className="flex gap-5 justify-center items-center">
             <h1